requirements.txt
# Core dependencies
# Note: pandas, numpy, scipy are not currently used in the codebase

# Web dashboard and visualization
dash>=2.14.0
plotly>=5.15.0

# Data handling and utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0

# Optimization libraries
pulp>=2.7.0  # For MILP optimization in optimal scheduler

# Optional: Advanced analytics (uncomment as needed)
# scikit-learn>=1.3.0  # For machine learning features
# networkx>=3.1  # For dependency graph analysis

